Move request models to the corresponding folder
