+++ lib/presentation/auth/change_password/request/bloc/change_password_request_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_template/data/auth/model/auth_data_error.dart';
+import 'package:flutter_template/data/auth/model/forgot_password_request/forgot_password_request_request_model.dart';
 import 'package:flutter_template/data/auth/repository/auth_repository.dart';
-import 'package:flutter_template/data/auth/service/remote/model/forgot_password_request/forgot_password_request_request_model.dart';
 import 'package:flutter_template/presentation/auth/change_password/request/bloc/change_password_request_error.dart';
 import 'package:flutter_template/presentation/auth/change_password/request/bloc/change_password_request_state.dart';
 import 'package:flutter_template/util/mixins/resettable_bloc_mixin.dart';
