+++ lib/presentation/auth/otp_verification/widgets/otp_resend_button/bloc/otp_resend_cubit.dart
@@ -20,7 +20,11 @@ class OtpResendCubit extends Cubit<OtpResendState> {
     Function(Object)? onError,
   }) async {
     try {
-      emit(const OtpResendStateLoading());
+      emit(
+        OtpResendStateLoading(
+          countdownSeconds: state.countdownSeconds,
+        ),
+      );
 
       await onResend();
